+++ Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
 [assembly: AssemblyTitle("Encrypted Email Bridge")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("Bridge between two or more locations on the Internet to send / receive files, using an email EmailHost / account.")]
-[assembly: AssemblyCompany("YucelG")]
+[assembly: AssemblyDescription("Bridge between two or more locations on the Internet to send / receive files, using an email EmailHost / account.")]
+[assembly: AssemblyConfiguration("retail")]
+[assembly: AssemblyCompany("GaziYucel")]
 [assembly: AssemblyProduct("EncryptedEmailBridge")]
-[assembly: AssemblyCopyright("Copyright YucelG")]
+[assembly: AssemblyCopyright("Copyright GaziYucel")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -31,5 +31,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.6.0.0")]
-[assembly: AssemblyFileVersion("1.6.0.0")]
+[assembly: AssemblyVersion("1.7.0.0")]
+[assembly: AssemblyFileVersion("1.7.0.0")]